Before : 

```
➜  php-redis-protocol (master) ✗ php example/perf.php 1000000
benchmarking 1000000 messages (chunksize of 4096 bytes)
2s for serialization
11.836s for parsing
➜  php-redis-protocol (master) ✗ php example/perf.php 1000000
benchmarking 1000000 messages (chunksize of 4096 bytes)
2.162s for serialization
11.803s for parsing
```

After : 

```
➜  php-redis-protocol (master) ✗ php example/perf.php 1000000
benchmarking 1000000 messages (chunksize of 4096 bytes)
2.022s for serialization
11.211s for parsing
➜  php-redis-protocol (master) ✗ php example/perf.php 1000000
benchmarking 1000000 messages (chunksize of 4096 bytes)
2.047s for serialization
11.109s for parsing
```