@clue 2015-09-09
Thanks for filing this pr @pborreli! Even if I don't expect this have a high impact, I still think it makes sense to address this :+1: